refactor: Change constants folder
