+++ tsconfig.json
@@ -20,7 +20,6 @@
          "@lib/*": ["lib/*"],
          "@features/*": ["features/*"],
          "@components/*": ["components/*"],
-         "@constants": ["constants"],
          "testUtils": ["./test/utils/testUtils"]
       },
       "types": ["cypress", "@testing-library/cypress"]
